+++ driver/bk4819.h
@@ -105,7 +105,7 @@ void     BK4819_DisableVox(void);
 void     BK4819_DisableDTMF(void);
 void     BK4819_EnableDTMF(void);
 void     BK4819_PlayTone(uint16_t Frequency, bool bTuningGainSwitch);
-void     BK4819_PlaySingleTone(const unsigned int tone_Hz, const unsigned int delay, const unsigned int level, const bool play_speaker);
+void     BK4819_PlaySingleTone(const uint16_t tone_Hz, const uint16_t delay, const uint16_t level, const bool play_speaker);
 void     BK4819_EnterTxMute(void);
 void     BK4819_ExitTxMute(void);
 void     BK4819_Sleep(void);
@@ -129,7 +129,7 @@ void     BK4819_EnableTXLink(void);
 void     BK4819_PlayDTMF(char Code);
 void     BK4819_PlayDTMFString(const char *pString, bool bDelayFirst, uint16_t FirstCodePersistTime, uint16_t HashCodePersistTime, uint16_t CodePersistTime, uint16_t CodeInternalTime);
 
-void     BK4819_TransmitTone(bool bLocalLoopback, uint32_t Frequency);
+void     BK4819_TransmitTone(bool bLocalLoopback, uint16_t Frequency);
 
 void     BK4819_GenTail(uint8_t Tail);
 void     BK4819_EnableCDCSS(void);
@@ -168,7 +168,7 @@ void     BK4819_PlayRoger();
 void     BK4819_Enable_AfDac_DiscMode_TxDsp(void);
 
 void     BK4819_GetVoxAmp(uint16_t *pResult);
-void     BK4819_SetScrambleFrequencyControlWord(uint32_t Frequency);
+void     BK4819_SetScrambleFrequencyControlWord(uint16_t Frequency);
 void     BK4819_PlayDTMFEx(bool bLocalLoopback, char Code);
 
 #endif
