+++ external/FreeRTOS/timers.c
@@ -453,7 +453,7 @@
             }
             else
             {
-                pxTimer->ucStatus &= ~tmrSTATUS_IS_AUTORELOAD;
+                pxTimer->ucStatus &= (uint8_t)~tmrSTATUS_IS_AUTORELOAD;
             }
         }
         taskEXIT_CRITICAL();
@@ -539,7 +539,7 @@
         }
         else
         {
-            pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+            pxTimer->ucStatus &= (uint8_t)~tmrSTATUS_IS_ACTIVE;
             mtCOVERAGE_TEST_MARKER();
         }
 
@@ -837,7 +837,7 @@
                     case tmrCOMMAND_STOP:
                     case tmrCOMMAND_STOP_FROM_ISR:
                         /* The timer has already been removed from the active list. */
-                        pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                        pxTimer->ucStatus &= (uint8_t)~tmrSTATUS_IS_ACTIVE;
                         break;
 
                     case tmrCOMMAND_CHANGE_PERIOD:
@@ -876,7 +876,7 @@
                                  * could not have been dynamically allocated. So there is
                                  * no need to free the memory - just mark the timer as
                                  * "not active". */
-                                pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                                pxTimer->ucStatus &= (uint8_t)~tmrSTATUS_IS_ACTIVE;
                             }
                         #endif /* configSUPPORT_DYNAMIC_ALLOCATION */
                         break;
