+++ radio/radio.h
@@ -155,7 +155,7 @@ extern VfoState_t     gVfoState;
 bool     RADIO_CheckValidChannel(uint16_t channel, bool checkScanList, uint8_t scanList);
 uint8_t  RADIO_FindNextChannel(uint8_t ChNum, int8_t Direction, bool bCheckScanList, uint8_t RadioNum);
 void     RADIO_InitInfo(VFO_Info_t *pInfo, const uint8_t ChannelSave, const uint32_t Frequency);
-void     RADIO_ConfigureChannel(const unsigned int VFO, const unsigned int configure);
+void     RADIO_ConfigureChannel(const uint8_t VFO, const uint8_t configure);
 void     RADIO_ConfigureSquelchAndOutputPower(VFO_Info_t *pInfo);
 void     RADIO_ApplyOffset(VFO_Info_t *pInfo);
 void     RADIO_SelectVfos(void);
